[#36] feat: useIntersectionObserver 커스텀 훅으로 분리
